Update MWE
Updated a Minimal Working Example (MWE). Let's just stick to plain old ROS for now. IGNORE ALL OTHER BRANCHES; THEY ARE DEFUNCT

Given the difficulties with ROS2 Rolling/Humble and the existing Moveit2 Move Group Interface I decided to deprecate the package to ROS Noetic; I did not go all the way down to Melodic since it is EOL right now and we need some sort of recent integration with Unity and the Audio package like I mentioned earlier.

I wasn't able to get the UR5 working with Noetic because the developers of the link-attacher plugins and the vacuum gripper services with the UR5 did not update their packages to work with Noetic. I have opted to use the Franka Emika Panda robot instead with plain RViz doing the rest for us instead of having a camera and all that.

I have added the following in the catkin_ws workspace folder:
- franka_description: URDF files and meshes of Franka Emika robots for RViz
- frankapanda_moveit_config: Custom moveit configuration of the Franka Emika Panda robot
- moveit_tutorials/move_group_python_interface: Repurposed moveit tutorial using the move group python interface (can't be bothered to rename and fix itty-bitty issues with the CMake/package right now)

@Adhithiyan03 You can add a subscriber for your ROS topic in the main function in the move_group_python_interface script file and link the data to the box_position variable that's taking care of the pick-placing.

Signed-off-by: Neehal <nasharrma@wpi.edu>