Update MWE, RViz Scenes
Cleaned up some code and added RViz scenes for the robot to use. There are two variants of the workspace that we will use for user studies. The first is a simple workspace with only the robot and the bins. The second is a more complex workspace with the robot, bins, and walls. The walls are used to block the robot's view of the bins. This is to test the user's ability to perceive the robot's ability to plan around obstacles.

Signed-off-by: Neehal <nasharrma@wpi.edu>